refactor(client): consolidate HTTP/1 request services
